Add syntax switching / pretty printing feature
This introduces a new concept of a `FinalResult`, that unifies a
CompilationResult.t and ConvertionResult.t. This makes sure we are
not getting out of sync with the error messages in the error pane.

Also it makes it easier to separate between pure syntax parsing results
and actual compilation results.