Update safari favicon (#281)
IIRC safari does a heuristics on the favicon png to see if the edges are filled colors. If so, display at regular size. If not (like in our case, where we already have rounded borders in the icon), then display at smaller size

We want the full sized icon. So this icon fills the edges.