Don't wipe playground url on load; update on copy to clipboard
Wiping the playground url turned out to be a less desirable UX than we
thought. Sometime tabs get backgrounded then ejected, so reloading those
with the default url gives us to the default snippet instead of the one
we loaded. Same for when we undo a closed tab.

I've removed that url wipe logic. I've also added the logic to update
the url on copy to clipboard, because why not. Maybe an incomplete UX,
but this beats updating the url every other second while typing, which
would be jarring