Clean up blog post data structure
This exemplifies our "static, catch things earlier" philosophy. The data shape and values of the blog authors are known in advance, always (and it's a good idea do have it so. I mean... what kind of better UX wants unknown runtime blog authors and all). So we get to remove whole chunks of optionalities here.