Highlighted UTF-8 / unicode compatibility gotchas
I had skimmed through the docs, but when I ran into the issue with `""` strings blowing up with UTF-8 contents (`Ü`, `♥`) I had no clue what I could do about this. Gladly some helpful people on the ReasonML Discord helped out with it, but before that I had spent quite a while searching for answers on both the docs and search engines.

The docs briefly mention the issue casually in middle of a title called "String interpolation" - not great. The only mention of "utf-8" on that page takes you to a section which says `Note: Char doesn't support Unicode or UTF-8 and is therefore not recommended.` - making it seem unicode is not supported at all.

Several of the top results on search engines seem to also indicate that OCaml has no support for unicode and so ReasonML/ReScript also cannot support it.

This clear highlight at the top of the page should at least help with identifying a common critical issue which a lot of users are guaranteed to run into.